Move part of Update to Core/Update
