+++ fragnix.cabal
@@ -46,6 +46,7 @@ library
                        Fragnix.Utils,
                        Fragnix.Update
                        Fragnix.Core.Slice
+                       Fragnix.Core.Update
   hs-source-dirs:      src
   ghc-options:         -O2 -Wall
